+++ src/plone/app/mosaic/tests/test_setup.py
@@ -1,8 +1,14 @@
 # -*- coding: utf-8 -*-
+from Products.CMFCore.utils import getToolByName
 from plone.app.mosaic.testing import PLONE_APP_MOSAIC_INTEGRATION
 from plone.browserlayer.utils import registered_layers
 from plone.registry.interfaces import IRegistry
 from zope.component import getUtility
+try:
+    from Products.CMFPlone.utils import get_installer
+except ImportError:
+    # BBB for Plone 5.0 and lower.
+    get_installer = None
 
 import unittest
 
@@ -26,8 +32,13 @@ class InstallTestCase(unittest.TestCase):
         self.portal = self.layer['portal']
 
     def test_installed(self):
-        qi = self.portal['portal_quickinstaller']
-        self.assertTrue(qi.isProductInstalled(PROJECTNAME))
+        if get_installer is None:
+            qi = getToolByName(self.portal, 'portal_quickinstaller')
+            installed = qi.isProductInstalled(PROJECTNAME)
+        else:
+            qi = get_installer(self.portal)
+            installed = qi.is_product_installed(PROJECTNAME)
+        self.assertTrue(installed)
 
     def test_addon_layer(self):
         layers = [l.getName() for l in registered_layers()]
@@ -56,11 +67,21 @@ class UninstallTestCase(unittest.TestCase):
 
     def setUp(self):
         self.portal = self.layer['portal']
-        self.qi = self.portal['portal_quickinstaller']
-        self.qi.uninstallProducts(products=[PROJECTNAME])
+        if get_installer is None:
+            qi = getToolByName(self.portal, 'portal_quickinstaller')
+            qi.uninstallProducts(products=[PROJECTNAME])
+        else:
+            qi = get_installer(self.portal)
+            qi.uninstall_product(PROJECTNAME)
 
     def test_uninstalled(self):
-        self.assertFalse(self.qi.isProductInstalled(PROJECTNAME))
+        if get_installer is None:
+            qi = getToolByName(self.portal, 'portal_quickinstaller')
+            installed = qi.isProductInstalled(PROJECTNAME)
+        else:
+            qi = get_installer(self.portal)
+            installed = qi.is_product_installed(PROJECTNAME)
+        self.assertFalse(installed)
 
     def test_addon_layer_removed(self):
         layers = [l.getName() for l in registered_layers()]